# api/openapi.yaml
openapi: 3.0.3
info:
  title: EDN Core API
  version: 1.0.0
  description: |
    Backend core para produtos digitais EDN — leads, checkout, pagamentos e assinaturas.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://dev.edn-core.app
    description: Cloud Run (dev)
  - url: https://edn-core.app
    description: Cloud Run (produção)
...
paths:
...
  /api/leads:
    post:
      summary: Submeter um lead
      description: |
        Recebe um novo lead de qualquer canal (web, mobile, parceiro).
        Requer o token de autenticação no header `Authorization`.
      tags: [Leads]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
            example:
              name: "Maria Silva"
              email: "maria@example.com"
              phone: "+5511999990000"
              source: "landing-page"
      responses:
        '201':
          description: Lead criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Lead duplicado (dedup window ativa)
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ── Organizations ────────────────────────────────────────────────────────────
...
  /v1/plans:
    get:
      summary: Listar planos ativos
      description: |
        Retorna os planos disponíveis para checkout do tenant.
        `organization_slug` e `tenant_slug` são obrigatórios. O backend resolve o tenant
        em runtime e consulta a coleção `versioned_plans`.
      tags: [Plans]
      parameters:
        - name: organization_slug
          in: query
          required: true
          description: Slug público da organização.
          schema:
            type: string
        - name: tenant_slug
          in: query
          required: true
          description: Slug público do tenant.
          schema:
            type: string
        - name: channel
          in: query
          required: false
          description: Canal comercial filtrado. O padrão operacional do checkout público é `web`.
          schema:
            type: string
            enum: [web, mobile, partner, all]
      responses:
        '200':
          description: Lista de planos
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plan'
              example:
                plans:
                  - slug: "pro"
                    name: "Plano Pro"
                    billing_cycle: "monthly"
                    price_cents: 9900
                    currency: "BRL"
        '404':
          description: Organização ou tenant não encontrados
        '422':
          description: Escopo inválido

  # ── Checkout ────────────────────────────────────────────────────────────────
  /v1/checkout/sessions:
    post:
      summary: Criar sessão de checkout
      description: |
        Cria um pedido e retorna a URL de checkout hospedada pela InfinitePay.
        O preço é sempre resolvido pelo backend a partir do plano — nunca aceito do cliente.
        O cliente deve enviar o header `Idempotency-Key` para garantir retry seguro.
        `organization_slug` e `tenant_slug` são obrigatórios; o backend resolve o tenant
        em runtime e seleciona o plano publicado em `versioned_plans`.
      tags: [Checkout]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Chave opaca do cliente para colapsar retries do mesmo checkout.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutRequest'
            example:
              organization_slug: "acme"
              tenant_slug: "public"
              channel: "web"
              plan_slug: "pro"
              billing_cycle: "monthly"
              customer:
                name: "João Costa"
                email: "joao@example.com"
                document: "12345678900"
                phone: "+5511988880000"
      responses:
        '201':
          description: Sessão criada — redirecionar para checkout_url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCheckoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Chave de idempotencia em conflito ou checkout ainda em andamento
        '404':
          description: Organização, tenant ou plano não encontrados
        '422':
          description: Requisição inválida
  /v1/orders/{orderNSU}/status:
    get:
      summary: Consultar status de pedido
      description: |
        Retorna o status atual do pedido e, se pago, os dados da assinatura ativada.
      tags: [Checkout]
      parameters:
        - name: orderNSU
          in: path
          required: true
          description: NSU único gerado no momento do checkout
          schema:
            type: string
            pattern: '^[a-zA-Z0-9\-_]{1,64}$'
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        '200':
          description: Status do pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Pedido não encontrado

# ── Components ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token de autenticação da leads API (VIL_LEADS_AUTH_TOKEN)
    adminAuth:
      type: http
      scheme: bearer
      description: Token de administração para organizations/tenants (VIL_HTTP_ADMIN_AUTH_TOKEN)
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Token ausente ou inválido
    TooManyRequests:
      description: Rate limit excedido
  schemas:
...
    CreateLeadRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        phone:
          type: string
          description: Número no formato E.164 ou brasileiro (10/11 dígitos)
          example: "+5511999990000"
        source:
          type: string
          maxLength: 80
          example: "landing-page"
    LeadResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Plans ────────────────────────────────────────────────────────────────
    Plan:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        billing_cycle:
          type: string
          enum: [monthly, annual]
        price_cents:
          type: integer
          description: Preço em centavos (BRL)
        currency:
          type: string
          example: BRL
        active:
          type: boolean

    # ── Checkout ─────────────────────────────────────────────────────────────
    CreateCheckoutRequest:
      type: object
      required: [organization_slug, tenant_slug, plan_slug, billing_cycle, customer]
      properties:
        organization_slug:
          type: string
          description: Slug público da organização.
        tenant_slug:
          type: string
          description: Slug público do tenant.
        channel:
          type: string
          enum: [web, mobile, partner, all]
          description: Canal comercial do checkout. Quando omitido, o backend assume `web`.
        plan_slug:
          type: string
        billing_cycle:
          type: string
          enum: [monthly, annual]
        customer:
          $ref: '#/components/schemas/Customer'
    Customer:
      type: object
      required: [name, email, phone, document]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          description: Telefone em E.164 (+5511...) ou 10/11 dígitos brasileiros
          example: "+5511987654321"
        document:
          type: string
          description: CPF — 11 dígitos com ou sem pontuação (000.000.000-00 ou 00000000000)
          example: "529.982.247-25"
    CreateCheckoutResponse:
      type: object
      properties:
        order_nsu:
          type: string
          description: Identificador único do pedido para consulta de status
        status:
          type: string
          enum: [created, checkout_created, pending, paid, failed, expired, pending_review]
        checkout_url:
          type: string
          format: uri
          description: URL de checkout da InfinitePay para redirecionar o cliente
        expires_at:
          type: string
          format: date-time
    OrderStatusResponse:
      type: object
      properties:
        order_nsu:
          type: string
        status:
          type: string
          enum: [created, checkout_created, pending, paid, failed, expired, pending_review]
        subscription_status:
          type: string
          enum: [pending_activation, active, suspended, canceled, expired]
        receipt_url:
          type: string
          format: uri
        plan_slug:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string